Issues #70 (#73)
* updated readme.md

* loose ends

* aptContact controller

* re-using aptList and Utils

* contacts screen

* edit contact screen

* Suggested text

* more informative contact screen

* whatsapp button widget BUGGED

* calendar screen changing months

* better snackbar

* adjusts to idle periods

* ammend: idle periods adjusts

* authorization added

* authorized api calls

* removing token expiration

* reading businessId from token. no tests yet

* string case-normalization

* normalizing idlePeriod.Name, adding [required] to apt prices

* renaming fields with caps-lock, and matching class counterparts fields

* cleaning models documentation comments

* tryParsing some inputs in dart. maxPrice? filter in GET

* cleaning controller documentation

* [foreignkey]

* tryna fix businessweek

* removing BusinessWeek

* remove postgre, use mysql

* updated README.md for MySql

* issues-#70 client_sdk

* fix: flutter build errors from new client generation

* removing warnings and commercial-hours from flutter

* removing required from prices in apts .cs files

* fix: calendar controller

* refresh indicator at apt screens

* businessConfig updating except services. trimming strings

* business services being updated

* fix: services input

* fix: edit schedule service

* removing un-implemented buttons

* ammend: ops, missed one

* order buttons ordering

* fix: looping login

* failed price fix

* flutter prices

* prices aint null anymore

* abreviation

* deactivate and delete account

* fix: customer record cpf bug

* bad business config snack

* get idle periods endpoint

* client-sdk get idle periods

* basic create idle period

* delete idle period via calendar

* punctuations

* added period name

* removed 'change telephone' screen

* too many services - backend

* subscriptions

* removing phoneNumber index

* customer phonenumber required servercontext

* subscription controller

* subscription workflow. Flow renamed to visit at frontend

* subscription migration with FIX and client-sdk

* fix: dont add subscription to pre-paid business

* register screen using cnpj, phone input fields

* name input field at register screen

* removing some TODOs

* removing more comments

* removing misteriously unnecessary code at aptlogcontroller

---------

+++ FixFlow.Server/Controllers/IdlePeriodController.cs
@@ -6,10 +6,10 @@ using Server.Models;
 using Server.Models.Erros;
 using Server.Models.Utils;
 
-namespace FixFlow.Server.Controllers.Users;
+namespace FixFlow.Server.Controllers;
 
 /// <summary>
-/// Controller class for Creating, Extending, Shrinking and Deleting Idle Periods
+/// Controller class for Creating and Deleting Idle Periods
 /// </summary>
 [ApiController]
 [Route(Common.api_v1 + nameof(IdlePeriod))]
@@ -24,53 +24,47 @@ public class IdlePeriodController : ControllerBase {
 	}
 
 	/// <summary>
-	/// Returns all Idle Periods that contain the given date
-	/// </summary>
-	[ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IdlePeriod[]))]
-	[HttpGet]
-	public async Task<IActionResult> GetIdlePeriodsAtDate(string businessId, DateTime date) {
-
-		var business = _context.Business.Find(businessId);
-		if (business == null) {
-			return BadRequest(NotExistErrors.Business);
-		}
-
-		var idlePeriods = await _context.IdlePeriods.Where(x => x.BusinessId == businessId)
-								.Where(x => x.start <= date && x.finish >= date)
-								.ToArrayAsync();
-
-		return Ok(idlePeriods);
-	}
-
-	/// <summary>
 	/// Creates an Idle period
 	/// </summary>
-	/// <remarks>
-	/// Idle Periods are allowed to overlap
-	/// </remarks>
 	[ProducesResponseType(StatusCodes.Status201Created, Type = typeof(IdlePeriod))]
 	[ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(string))]
 	[HttpPost]
 	public async Task<IActionResult> CreateIdlePeriod([FromBody] IdlePeriod idlePeriod) {
 
-		if (_context.Business.Find(idlePeriod.BusinessId) == null) {
-			return BadRequest(NotExistErrors.Business);
-		}
+		string businessId = User.Claims.First(c => c.Type == "businessId")?.Value!;
 
-		if (idlePeriod.start <= DateTime.UtcNow.AddMinutes(-5) || idlePeriod.finish <= DateTime.UtcNow) {
-			return BadRequest(ValidatorErrors.IdlePeriodHasPassed);
+		if (idlePeriod.BusinessId != businessId) {
+			return BadRequest(ValidatorErrors.BadIdlePeriodOwnership);
 		}
 
 		idlePeriod.Id = Guid.NewGuid().ToString();
-		_context.IdlePeriods.Add(idlePeriod);
+		idlePeriod.Name = StringUtils.PhraseCaseNormalizer(idlePeriod.Name)!;
 
+		_context.IdlePeriods.Add(idlePeriod);
 		await _context.SaveChangesAsync();
 
-		return Ok(idlePeriod);
+		return CreatedAtAction(nameof(CreateIdlePeriod), idlePeriod);
 	}
 
 	/// <summary>
-	/// Removes Idle days
+	/// Gets Idle Periods owned by the Company that start and end within a given time-period
+	/// </summary>
+	[ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IdlePeriod))]
+	[HttpGet]
+	public async Task<IActionResult> GetIdlePeriod(DateTime startDate, DateTime finishDate) {
+
+		string businessId = User.Claims.First(c => c.Type == "businessId")?.Value!;
+
+		IdlePeriod[] periods = await _context.IdlePeriods
+		.Where(i => i.Start >= startDate)
+		.Where(i => i.Finish <= finishDate)
+		.ToArrayAsync();
+
+		return Ok(periods);
+	}
+
+	/// <summary>
+	/// Deletes an Idle Period
 	/// </summary>
 	[ProducesResponseType(StatusCodes.Status204NoContent)]
 	[ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(string))]
@@ -79,7 +73,12 @@ public class IdlePeriodController : ControllerBase {
 
 		var idlePeriod = _context.IdlePeriods.Find(idlePeriodId);
 		if (idlePeriod == null) {
-			return BadRequest(NotExistErrors.IdlePeriod);
+			return BadRequest(NotExistErrors.idlePeriod);
+		}
+
+		string businessId = User.Claims.First(c => c.Type == "businessId")?.Value!;
+		if (idlePeriod.BusinessId != businessId) {
+			return Unauthorized(ValidatorErrors.BadIdlePeriodOwnership);
 		}
 
 		_context.IdlePeriods.Remove(idlePeriod);
